Centralize S2699 framework detection
